Added Wicherts et al 2006 Email author != success
Although it deals with another discipline, namely psychological research, this is a must-cite paper IMO. It proves that any kind of 'email the author' for the data system will NOT work and so is not good enough.
Wicherts, J. M., Borsboom, D., Kats, J., and Molenaar, D. 2006. The poor availability of psychological research data for reanalysis. American Psychologist 61:726-728. http://dx.doi.org/10.1037/0003-066X.61.7.726