Sentence deletion
"By making dataset public and citeable, the contribution of data
will become less and less of a criteria for authorship."

I think this sentence could be all too easily mis-understood. I infer you're referring to authorship traded for re-use of 'already published-on' data here. In that sense I agree with the statement to some extent, although it's still a bit of a strong and unevidenced assertion. However, the trouble is - that sentence could also be inferred to mean the contribution of *newly* collected data. E.g. someone works hard to collect data for weeks from a very isolated, remote, expensive to reach location. Contributing significant *new* data to paper could/ sometimes should still lead to authorship in my opinion and thus I've deleted this sentence to avoid confusion. Either delete or rewrite to make clearer what you mean in the context of authorship.

What authorship currently is (a mess) and what authorship should be (hotly debated) are difficult to pin down. Perhaps it might be worth pointing out that data can have completely different 'authorship' to the authorship of a paper that may use or refer to that data. Repositories like Dryad already allow for this e.g. I was a middle author on a particular paper, but I was 1st author of the associated Dryad data package :)